feat: first implementation for supporting building on linux systems
