+++ api/v0/config/config.proto
@@ -8,7 +8,7 @@ service ConfigService {
 }
 
 enum ConfigPackageSourceKind {
-    UNKNOWN = 0;
+    UNKNOWN_SOURCE = 0;
     LOCAL = 1;
     HTTP = 2;
     GIT = 3;
@@ -19,11 +19,18 @@ message ConfigPackageOutput {
     string hash = 2;
 }
 
+enum ConfigPackageBuildSystem {
+    UNKNOWN_SYSTEM = 0;
+    AARCH64_LINUX = 1;
+    AARCH64_MACOS = 2;
+}
+
 message ConfigPackageBuild {
     bool sandbox = 1;
     map<string, string> environment = 2;
     repeated ConfigPackageOutput packages = 3;
     string script = 4;
+    ConfigPackageBuildSystem system = 5;
 }
 
 message ConfigPackageSource {
